Rename VoiceController to Voice::OtpController
**Why**: Follow more rails-y naming conventions
