+++ .slim-lint.yml
@@ -3,7 +3,7 @@ linters:
     max: 100
   TagCase:
     exclude:
-     - 'app/views/voice/otp.xml.slim'
+     - 'app/views/voice/otp/show.xml.slim'
   RuboCop:
     enabled: true
     # slim-lint ignores all of these cops by default, and we need to copy them
